+++ src/components/TeacherComponents/UnitList/UnitList.tsx
@@ -1,5 +1,6 @@
-import React, { FC, MouseEvent } from "react";
+import React, { FC, MouseEvent, useState } from "react";
 import { NextRouter, useRouter } from "next/router";
+import { useFeatureFlagEnabled } from "posthog-js/react";
 import {
   OakFlex,
   OakUnitsContainer,
@@ -8,12 +9,14 @@ import {
   OakPagination,
   OakAnchorTarget,
   OakBox,
+  OakInlineBanner,
 } from "@oaknational/oak-components";
 
 import { UnitOption } from "../UnitListOptionalityCard/UnitListOptionalityCard";
 import { getSubjectPhaseSlug } from "../helpers/getSubjectPhaseSlug";
 
 import { getPageItems, getProgrammeFactors } from "./helpers";
+import { UnitListLegacyBanner } from "./UnitListLegacyBanner";
 
 import {
   UnitListItemProps,
@@ -82,14 +85,21 @@ const getOptionalityUnits = (
         router.push(target.href);
       }
     };
+    const lessonCount = getUnitLessonCount({
+      lessonCount: unitOption.lessonCount,
+      expiredLessonCount: unitOption.expiredLessonCount,
+      unpublishedLessonCount: unitOption.unpublishedLessonCount,
+    });
+
     return {
       title: unitOption.title,
+      slug: unitOption.slug,
       href: resolveOakHref({
         page: "lesson-index",
         unitSlug: unitOption.slug,
         programmeSlug: unitOption.programmeSlug,
       }),
-      lessonCount: unitOption.lessonCount || 0,
+      lessonCount,
       onClick: handleClick,
     };
   });
@@ -151,7 +161,7 @@ const UnitList: FC<UnitListProps> = (props) => {
   const linkSubject = subjectParent
     ? convertSubjectToSlug(subjectParent)
     : subjectSlug;
-  const { currentPage, pageSize, firstItemRef, paginationRoute } =
+  const { currentPage, pageSize, firstItemRef, paginationRoute, onPageChange } =
     paginationProps;
   const router = useRouter();
   const category = router.query["category"]?.toString();
@@ -179,6 +189,20 @@ const UnitList: FC<UnitListProps> = (props) => {
   const indexOfFirstLegacyUnit = units
     .map((u) => isSlugLegacy(u[0]!.programmeSlug))
     .indexOf(true);
+
+  // stub implementation of saving
+  const isSaveEnabled = useFeatureFlagEnabled("teacher-save-units");
+  const [savedUnitsForUser, setSavedUnitsForUser] = useState<string[]>([]);
+
+  const onSave = (unitSlug: string) => {
+    const newSavedUnits = savedUnitsForUser.includes(unitSlug)
+      ? savedUnitsForUser.filter((u) => u !== unitSlug)
+      : [...savedUnitsForUser, unitSlug];
+    setSavedUnitsForUser(newSavedUnits);
+  };
+
+  const hasNewAndLegacyUnits: boolean =
+    !!phaseSlug && !!newPageItems.length && !!legacyPageItems.length;
 
   const getUnitCards = (
     pageItems: CurrentPageItemsProps[] | SpecialistUnit[][],
@@ -238,6 +262,8 @@ const UnitList: FC<UnitListProps> = (props) => {
               : null
           }
           optionalityUnits={getOptionalityUnits(item, onClick, router)}
+          onSave={isSaveEnabled ? onSave : undefined}
+          getIsSaved={(unitSlug) => savedUnitsForUser.includes(unitSlug)}
         />
       ) : (
         item.map((unitOption) => {
@@ -256,6 +282,7 @@ const UnitList: FC<UnitListProps> = (props) => {
             expiredLessonCount: unitOption.expiredLessonCount,
             unpublishedLessonCount: unitOption.unpublishedLessonCount,
           });
+
           return (
             <OakUnitListItem
               {...props}
@@ -283,6 +310,8 @@ const UnitList: FC<UnitListProps> = (props) => {
                 unitSlug: unitOption.slug,
                 programmeSlug: unitOption.programmeSlug,
               })}
+              onSave={isSaveEnabled ? () => onSave(unitOption.slug) : undefined}
+              isSaved={savedUnitsForUser.includes(unitOption.slug)}
             />
           );
         })
@@ -321,6 +350,14 @@ const UnitList: FC<UnitListProps> = (props) => {
     legacyPageItems.length && keyStageSlug && phaseSlug ? (
       <OakUnitsContainer
         isLegacy={true}
+        banner={
+          <UnitListLegacyBanner
+            userType={"teacher"}
+            hasNewUnits={hasNewAndLegacyUnits}
+            allLegacyUnits={legacyPageItems}
+            onButtonClick={() => onPageChange(1)}
+          />
+        }
         subject={subjectSlug}
         phase={phaseSlug}
         curriculumHref={resolveOakHref({
@@ -362,7 +399,16 @@ const UnitList: FC<UnitListProps> = (props) => {
           unitCards={getUnitCards(swimmingPageItems)}
           isCustomUnit={true}
           customHeadingText={title}
-          bannerText="Swimming and water safety lessons should be selected based on the ability and experience of your pupils."
+          banner={
+            <OakInlineBanner
+              isOpen={true}
+              message={
+                "Swimming and water safety lessons should be selected based on the ability and experience of your pupils."
+              }
+              type="neutral"
+              $width={"100%"}
+            />
+          }
         />
       );
     } else return null;
@@ -381,6 +427,14 @@ const UnitList: FC<UnitListProps> = (props) => {
         ) : (
           <OakUnitsContainer
             isLegacy={true}
+            banner={
+              <UnitListLegacyBanner
+                userType={"teacher"}
+                hasNewUnits={hasNewAndLegacyUnits}
+                allLegacyUnits={legacyPageItems}
+                onButtonClick={() => onPageChange(1)}
+              />
+            }
             subject={subjectSlug}
             phase={"Specialist and therapies"}
             curriculumHref={`${resolveOakHref({