+++ src/components/TeacherViews/Search/helpers/index.ts
@@ -1,4 +1,8 @@
-import { KeyStageTitleValueType } from "@/browser-lib/avo/Avo";
+import {
+  FilterTypeValueType,
+  KeyStageTitleValueType,
+  SearchFilterModifiedProperties,
+} from "@/browser-lib/avo/Avo";
 
 export const convertUnitSlugToTitle = (unitSlug: string) => {
   const lastHyphenIndex = unitSlug.lastIndexOf("-");
@@ -26,3 +30,30 @@ export const isKeyStageTitleValueType = (
     value === "Key stage 4"
   );
 };
+
+export type TrackSearchModifiedProps = {
+  checked: boolean;
+  filterType: FilterTypeValueType;
+  filterValue: string;
+};
+
+export const trackSearchModified =
+  (
+    query: string,
+    searchFilterModified: (props: SearchFilterModifiedProperties) => void,
+  ) =>
+  ({ checked, filterType, filterValue }: TrackSearchModifiedProps) => {
+    const filterModificationType = checked ? "remove" : "add";
+    searchFilterModified({
+      platform: "owa",
+      product: "teacher lesson resources",
+      engagementIntent: "refine",
+      componentType: "filter_link",
+      eventVersion: "2.0.0",
+      analyticsUseCase: "Teacher",
+      filterModificationType,
+      filterType,
+      filterValue,
+      searchTerm: query,
+    });
+  };