+++ src/styles/GlobalStyle.ts
@@ -4,6 +4,7 @@ import { createGlobalStyle, css } from "styled-components";
 import resetStyles from "./reset.styles";
 import oakStyles from "./oak.styles";
 import posthogHackStyles from "./posthog-hack.styles";
+import gleapHackStyles from "./gleap-hack.styles";
 
 type GlobalStyleProps = {
   fontFamily: string;
@@ -13,6 +14,7 @@ const GlobalStyle = createGlobalStyle<GlobalStyleProps>`
   ${resetStyles}
   ${oakStyles}
   ${posthogHackStyles}
+  ${gleapHackStyles}
   ${(props) => {
     return css`
       :root {